Restructure repo into monorepo style: add unit-specific logos and reservation views
