+++ app/Http/Controllers/CampaignControllerRCD.php
@@ -17,7 +17,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
-class CampaignController extends Controller
+class CampaignControllerRCD extends Controller
 {
     use UserLogsActivity;
