+++ composer.json
@@ -47,7 +47,10 @@
             "App\\": "app/",
             "Database\\Factories\\": "database/factories/",
             "Database\\Seeders\\": "database/seeders/"
-        }
+        },
+        "files": [
+            "app/Helpers/UnitHelper.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
